+++ decidim-budgets/lib/decidim/budgets/component.rb
@@ -39,9 +39,13 @@ Decidim.register_component(:budgets) do |component|
 
   component.settings(:global) do |settings|
     settings.attribute :projects_per_page, type: :integer, default: 12
+    settings.attribute :vote_per_budget, type: :boolean
     settings.attribute :total_budget, type: :integer, default: 100_000_000
     settings.attribute :vote_threshold_percent, type: :integer, default: 70
+    settings.attribute :vote_per_project, type: :boolean
+    settings.attribute :total_projects, type: :integer, default: 5
     settings.attribute :comments_enabled, type: :boolean, default: true
+    settings.attribute :geocoding_enabled, type: :boolean
     settings.attribute :announcement, type: :text, translated: true, editor: true
   end